requirements.txt
tensorflow>=2.13.0
onnxruntime>=1.16.0
mtcnn>=0.1.1
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
aiohttp>=3.9.0
pydantic>=2.5.0
python-multipart>=0.0.6
requests>=2.31.0
typing-extensions>=4.8.0
scikit-learn>=1.3.0 